Implement JWT with Role and Refresh Token.

+++ ComputerPartsShop.Infrastructure/Repositories/UserPaymentSystemRepository.cs
@@ -44,14 +44,14 @@ namespace ComputerPartsShop.Infrastructure
 			}
 		}
 
-		public async Task<UserPaymentSystem> GetAsync(Guid id, CancellationToken ct)
+		public async Task<UserPaymentSystem> GetAsync(Guid id, string username, CancellationToken ct)
 		{
 			var query = "SELECT UserPaymentSystem.ID, UserPaymentSystem.PaymentReference, ShopUser.Username, ShopUser.Email, PaymentProvider.Name, Payment.ID, Payment.OrderID, " +
 				"Payment.Total, Payment.Method, Payment.Status, Payment.PaymentStartAt, Payment.PaidAt FROM UserPaymentSystem " +
 				"LEFT JOIN ShopUser ON ShopUser.ID = UserPaymentSystem.UserID " +
 				"LEFT JOIN PaymentProvider ON PaymentProvider.ID = UserPaymentSystem.ProviderID " +
 				"LEFT JOIN Payment ON Payment.UserPaymentSystemID = UserPaymentSystem.ID " +
-				"WHERE UserPaymentSystem.ID = @ID";
+				"WHERE UserPaymentSystem.ID = @ID AND ShopUser.Username = @Username";
 
 			var userPaymentSystemDictionary = new Dictionary<Guid, UserPaymentSystem>();
 
@@ -77,7 +77,7 @@ namespace ComputerPartsShop.Infrastructure
 						}
 
 						return cps;
-					}, new { ID = id }, splitOn: "Username, Name, ID");
+					}, new { ID = id, Username = username }, splitOn: "Username, Name, ID");
 
 					return result.Distinct().FirstOrDefault();
 				}
@@ -123,16 +123,18 @@ namespace ComputerPartsShop.Infrastructure
 			}
 		}
 
-		public async Task<UserPaymentSystem> UpdateAsync(Guid id, UserPaymentSystem request, CancellationToken ct)
+		public async Task<int> UpdateAsync(Guid id, string username, UserPaymentSystem request, CancellationToken ct)
 		{
 			request.Id = id;
-			var query = "UPDATE UserPaymentSystem SET UserID = @UserID, ProviderID = @ProviderID, PaymentReference = @PaymentReference WHERE ID = @ID";
+			var query = "UPDATE UserPaymentSystem SET UserID = @UserID, ProviderID = @ProviderID, PaymentReference = @PaymentReference " +
+				"JOIN ShopUser ON ShopUser.ID = UserPaymentSystem.UserID WHERE UserPaymentSystem.ID = @ID AND Username = @Username";
 
 			var parameters = new DynamicParameters();
 			parameters.Add("ID", request.Id, DbType.Guid, ParameterDirection.Input);
 			parameters.Add("UserID", request.UserId, DbType.Guid, ParameterDirection.Input);
 			parameters.Add("ProviderID", request.ProviderId, DbType.Int32, ParameterDirection.Input);
 			parameters.Add("PaymentReference", request.PaymentReference, DbType.String, ParameterDirection.Input);
+			parameters.Add("Username", username, DbType.String, ParameterDirection.Input);
 
 			using (var connection = await _dbContext.CreateConnection())
 			{
@@ -140,10 +142,10 @@ namespace ComputerPartsShop.Infrastructure
 				{
 					try
 					{
-						await connection.ExecuteAsync(query, parameters, transaction);
+						var rowsAffected = await connection.ExecuteAsync(query, parameters, transaction);
 						transaction.Commit();
 
-						return request;
+						return rowsAffected;
 					}
 					catch (SqlException ex)
 					{
@@ -156,9 +158,9 @@ namespace ComputerPartsShop.Infrastructure
 			}
 		}
 
-		public async Task DeleteAsync(Guid id, CancellationToken ct)
+		public async Task<bool> DeleteAsync(Guid id, string username, CancellationToken ct)
 		{
-			var query = "DELETE FROM UserPaymentSystem WHERE ID = @Id";
+			var query = "DELETE FROM UserPaymentSystem JOIN ShopUser ON ShopUser.ID = UserPaymentSystem.UserID WHERE ID = @Id AND ShopUser.Username = @Username";
 
 			using (var connection = await _dbContext.CreateConnection())
 			{
@@ -166,8 +168,10 @@ namespace ComputerPartsShop.Infrastructure
 				{
 					try
 					{
-						await connection.ExecuteAsync(query, new { ID = id }, transaction);
+						int rowsAffected = await connection.ExecuteAsync(query, new { ID = id }, transaction);
 						transaction.Commit();
+
+						return rowsAffected > 0;
 					}
 					catch (SqlException ex)
 					{
